# src/main/resources/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
  "tags": [
    {
      "name": "basic-error-controller",
      "description": "Basic Error Controller"
    },
    {
      "name": "exam-controller",
      "description": "Exam Controller"
    },
    {
      "name": "web-mvc-links-handler",
      "description": "Web Mvc Links Handler"
    }
  ],
  "paths": {
    "/error": {
      "get": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingPUT",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingPOST",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingDELETE",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "options": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingOPTIONS",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "head": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingHEAD",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "patch": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingPATCH",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "trace": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingTRACE",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/exams": {
      "get": {
        "tags": [
          "exam-controller"
        ],
        "summary": "findAll",
        "operationId": "findAllUsingGET",
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "pageNo",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "302": {
            "description": "Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamDTORes"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "exam-controller"
        ],
        "summary": "create",
        "operationId": "createUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamDTOReq"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/exams/{id}": {
      "get": {
        "tags": [
          "exam-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "302": {
            "description": "Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDTORes"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "exam-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamDTOReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "exam-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/actuator": {
      "get": {
        "tags": [
          "web-mvc-links-handler"
        ],
        "summary": "links",
        "operationId": "linksUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              },
              "application/vnd.spring-boot.actuator.v3+json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              },
              "application/vnd.spring-boot.actuator.v2+json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerDTOReq": {
        "title": "AnswerDTOReq",
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "format": "byte"
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDTOReq"
          }
        }
      },
      "AnswerDTORes": {
        "title": "AnswerDTORes",
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "format": "byte"
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "new": {
            "type": "boolean"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDTORes"
          }
        }
      },
      "ChoiceDTOReq": {
        "title": "ChoiceDTOReq",
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "format": "byte"
          },
          "explanation": {
            "type": "string"
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "type": "string",
            "format": "byte"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDTOReq"
          },
          "weight": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ChoiceDTORes": {
        "title": "ChoiceDTORes",
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "format": "byte"
          },
          "explanation": {
            "type": "string"
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "type": "string",
            "format": "byte"
          },
          "new": {
            "type": "boolean"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDTORes"
          },
          "weight": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ExamDTOReq": {
        "title": "ExamDTOReq",
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "ended": {
            "type": "string",
            "format": "date-time"
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDTOReq"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "started": {
            "type": "string",
            "format": "date-time"
          },
          "student": {
            "$ref": "#/components/schemas/UserDTOReq"
          }
        }
      },
      "ExamDTORes": {
        "title": "ExamDTORes",
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "ended": {
            "type": "string",
            "format": "date-time"
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "new": {
            "type": "boolean"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDTORes"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "started": {
            "type": "string",
            "format": "date-time"
          },
          "student": {
            "$ref": "#/components/schemas/UserDTORes"
          }
        }
      },
      "Link": {
        "title": "Link",
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "ModelAndView": {
        "title": "ModelAndView",
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "model": {
            "type": "object"
          },
          "modelMap": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "reference": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "ALREADY_REPORTED",
              "BAD_GATEWAY",
              "BAD_REQUEST",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "CHECKPOINT",
              "CONFLICT",
              "CONTINUE",
              "CREATED",
              "DESTINATION_LOCKED",
              "EXPECTATION_FAILED",
              "FAILED_DEPENDENCY",
              "FORBIDDEN",
              "FOUND",
              "GATEWAY_TIMEOUT",
              "GONE",
              "HTTP_VERSION_NOT_SUPPORTED",
              "IM_USED",
              "INSUFFICIENT_SPACE_ON_RESOURCE",
              "INSUFFICIENT_STORAGE",
              "INTERNAL_SERVER_ERROR",
              "I_AM_A_TEAPOT",
              "LENGTH_REQUIRED",
              "LOCKED",
              "LOOP_DETECTED",
              "METHOD_FAILURE",
              "METHOD_NOT_ALLOWED",
              "MOVED_PERMANENTLY",
              "MOVED_TEMPORARILY",
              "MULTIPLE_CHOICES",
              "MULTI_STATUS",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NOT_ACCEPTABLE",
              "NOT_EXTENDED",
              "NOT_FOUND",
              "NOT_IMPLEMENTED",
              "NOT_MODIFIED",
              "NO_CONTENT",
              "OK",
              "PARTIAL_CONTENT",
              "PAYLOAD_TOO_LARGE",
              "PAYMENT_REQUIRED",
              "PERMANENT_REDIRECT",
              "PRECONDITION_FAILED",
              "PRECONDITION_REQUIRED",
              "PROCESSING",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_RANGE_NOT_SATISFIABLE",
              "REQUEST_ENTITY_TOO_LARGE",
              "REQUEST_HEADER_FIELDS_TOO_LARGE",
              "REQUEST_TIMEOUT",
              "REQUEST_URI_TOO_LONG",
              "RESET_CONTENT",
              "SEE_OTHER",
              "SERVICE_UNAVAILABLE",
              "SWITCHING_PROTOCOLS",
              "TEMPORARY_REDIRECT",
              "TOO_EARLY",
              "TOO_MANY_REQUESTS",
              "UNAUTHORIZED",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "UPGRADE_REQUIRED",
              "URI_TOO_LONG",
              "USE_PROXY",
              "VARIANT_ALSO_NEGOTIATES"
            ]
          },
          "view": {
            "$ref": "#/components/schemas/View"
          },
          "viewName": {
            "type": "string"
          }
        }
      },
      "QuestionDTOReq": {
        "title": "QuestionDTOReq",
        "type": "object",
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/AnswerDTOReq"
          },
          "body": {
            "type": "string"
          },
          "choices": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChoiceDTOReq"
            }
          },
          "exams": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamDTOReq"
            }
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionDTORes": {
        "title": "QuestionDTORes",
        "type": "object",
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/AnswerDTORes"
          },
          "body": {
            "type": "string"
          },
          "choices": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChoiceDTORes"
            }
          },
          "exams": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamDTORes"
            }
          },
          "header": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "new": {
            "type": "boolean"
          }
        }
      },
      "UserDTOReq": {
        "title": "UserDTOReq",
        "type": "object",
        "properties": {
          "activated": {
            "type": "boolean"
          },
          "activation": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "exams": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamDTOReq"
            }
          },
          "firstName": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "roles": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "session": {
            "type": "string"
          }
        }
      },
      "UserDTORes": {
        "title": "UserDTORes",
        "type": "object",
        "properties": {
          "activated": {
            "type": "boolean"
          },
          "activation": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "exams": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamDTORes"
            }
          },
          "firstName": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lastName": {
            "type": "string"
          },
          "new": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "roles": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "session": {
            "type": "string"
          }
        }
      },
      "View": {
        "title": "View",
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string"
          }
        }
      }
    }
  }
}